Switch arc- and mesa-fredreno ebuilds to mesa-freedreno checkout
This depends on manifest changes to switch the mesa-freedreno checkout
to the chromeos-freedreno branch.  Also, move mesa-freedreno to be a
regular cros_workon ebuild.  We'll use the standard UPSTREAM/FROMLIST
prefixes for downstream patches, but ideally we should be able to
rebase often and avoid those.

BUG=b:116870852
TEST=verify that emerge builds from mesa-freedreno branch and that
  ebuilds update properly when updating the chromeos-freedreno branch.
CQ-DEPEND=CL:1415837, CL:*795566

Change-Id: I60f6912fea5583c9c36916ddb4f47782386daaef
Reviewed-on: https://chromium-review.googlesource.com/1416851
Commit-Ready: Kristian H. Kristensen <hoegsberg@chromium.org>
Tested-by: Kristian H. Kristensen <hoegsberg@chromium.org>
Reviewed-by: Chad Versace <chadversary@chromium.org>
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>