chromeos-chrome: disable lld in non-internal builds
lld has a bug in how it emits unwinding info. Since Chrome sometimes
tries to unwind itself (for `StackTrace`s), this causes Chrome to crash.
chrome_internal builds explicitly disable the emission of said bugged
unwinding info, so continuing to use lld for those should be safe.

This is intended to be reverted when the relevant bug is fixed.

BUG=chromium:917504
TEST=`emerge-kevin` with and without chrome_internal; args.gn looked
     reasonable, and the expected warning was emitted.
     emerge-kevin chromeos-base/chromeos-chrome succeeded with:
     	- USE=
	- USE=chrome_internal
	- USE="gold -lld"

Change-Id: I7958310330cd41ddd78b98da0c1f1f83e0af41bb
Reviewed-on: https://chromium-review.googlesource.com/1405444
Commit-Ready: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>