rust: Support stripping patch.crates-io entries from Cargo.toml
This CL introduces the ability for Cargo.toml files of cros-rust targets
to contain:

    # ignore in ebuild {
    ...
    # }

in between which the contents are stripped out during ebuild.

We will use this to discard parts of [patch.crates-io] which should
apply to local developer builds but not to ebuilds. In particular, some
crates contained within the crosvm repository have their own ebuild
independent of the crosvm ebuild so that they are usable from outside of
crosvm. When a developer is building these crates locally with Cargo, we
would like to patch dependencies such that they point to sibling crates
within the crosvm repository. But when building via ebuild, the patches
must not exist because they violate our build sandbox.

The CROS_WORKON_OUTOFTREE_BUILD=1 flag must not be set in an ebuild that
relies on `ignore in ebuild`.

BUG=chromium:916921
TEST=emerge-amd64-generic crosvm, data_model, sys_util (with and without CL:1403887)

Change-Id: I6b8413d14005fa1ab64549b31a17a4344356e535
Reviewed-on: https://chromium-review.googlesource.com/1407464
Commit-Ready: David Tolnay <dtolnay@chromium.org>
Tested-by: David Tolnay <dtolnay@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>