iwlwifi_rescan: add iwlwifi module params
For now, we set this for a few boards directly in the kernel command
line, in the relevant overlays. But this is difficult and a bit
redundant, because we also have to pull in the
net-wireless/iwlwifi_rescan dependency. Also, iwlwifi is never built-in
(it's always a module), so we can rely on modprobe.d to set up the
parameter for us.

Make this easier to manage by specifying the parameter in only one
place.

BUG=chromium:922232
TEST=check /sys/module/iwlwifi/parameters/remove_when_gone on glimmer
     after installing this file

Change-Id: Ief6b54837405e0269c10094e61796510d518c408
Reviewed-on: https://chromium-review.googlesource.com/1412451
Commit-Ready: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Rajat Jain <rajatja@chromium.org>