target-chromium-os: add iwlwifi_rescan USE flag
Support for removing/rescanning the Intel WiFi device is provided by the
net-wireless/iwlwifi_rescan ebuild, as a workaround for some known
hardware mis-behavior. It can legitimately be enabled on any platform,
as its effects are only seen when the iwlwifi driver is present and a
known PCI ID is seen in the system. But we enable it selectively
(a) to save space
(b) to apply the workaround where we have best understood that it is
    needed (and have no chance of effecting a hardware change instead)

Currently, it's only available on 3 boards, but we plan to widen that
scope. Rather than adding new dependencies to BSP ebuilds all over the
place, I find it cleaner to control this via USE flag.

BUG=chromium:922232
TEST=precq

Change-Id: I2892059868aa656ee7fcd2b2f53465c54d45d390
Reviewed-on: https://chromium-review.googlesource.com/1413611
Commit-Ready: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Rajat Jain <rajatja@chromium.org>