dnsmasq: Addition of upstream dnsmasq patches
In some scenarios, whenever a station tries to connect to the router,
the DHCP-DORA process happens successfully and after few seconds the
station sends a DHCP-DECLINE message to the router.

In the process of allocating a address to a client, DNSMASQ will
perform a PING check on the tentative address and, based on the
PING result, it will offer the address to the client(if PING fails)
or it will check for the next address in the pool
(when PING is successful).

If there are any stations in the LAN network which block PING packets
from other devices in the LAN, then we end up in offering the same
IP address to different clients.

This CL adds an additional logic of skipping addresses equal to the
number of addresses rejected by clients in sequential-ip mode.

Design Doc link:
https://docs.google.com/document/d/1ZBtcA3cfUG5H0mE0ExhdUmeFoVhmP1uUCudp8DEdGuY/edit#

BUG=b:112286158
TEST=Tested the following scenario on router.
    - Before the start of the test, the max leased address is
      192.168.86.23.
    - Connected a station statically with the IP address 192.168.86.24
      and also has a iptable rule added to block the ICMP ECHO REQUEST
      packets and the rule is as follows:
      sudo iptables -I INPUT -p icmp --icmp-type echo-request -j DROP
    - Connected macbook with IP address configuration as DHCP.
      Initially it got assigned with 192.168.86.24 but macbook
      declined the address and the next time when it initiated DORA
      process it got assigned with the IP address 192.168.86.25.

Change-Id: Ib1b0e2debe74fa1d2256bd6b0ed74faeff762f80
Reviewed-on: https://chromium-review.googlesource.com/1414559
Commit-Ready: Anvesh Jangala <a.jangala@globaledgesoftware.corp-partner.google.com>
Tested-by: Anvesh Jangala <a.jangala@globaledgesoftware.corp-partner.google.com>
Reviewed-by: Raju Konduru <rkonduru@google.com>
Reviewed-by: Kevin Hayes <kevinhayes@google.com>
Reviewed-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Kishan Kunduru <kkunduru@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>