media-libs/libepoxy: use virtual/opengles instead of mesa as an RDEPEND
This library is needed by virglrendener, which in turn is needed by
crosvm with gpu support. Because mesa is a dependency before this CL, it
would fail to install on boards that use non-mesa GL drivers, such as
mali boards. This CL changes the DEPEND to use the virtual opengles
package to work with any board's GLES implementation.

BUG=chromium:837073
TEST=emerge-kevin libepoxy

Change-Id: Ibdc5bf8575e9e808f75f93f04057df8a270510a6
Reviewed-on: https://chromium-review.googlesource.com/1418939
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>