sys-boot/coreboot: Enable overrides for microcode
A recent change (CL:1404987) to coreboot requires us to select a config
option before being able to choose our own path to the microcode updates
we want to add to the image.

Add that option so we're not stuck with whatever we got from upstream
(or nothing at all!) when we ship a different option, usually for good
reason.

BUG=b:122854057
BRANCH=none
TEST=emerge-$board coreboot shows a different file size for
the cpu_microcode_blob.bin file in the final image. Observed on atlas.

Change-Id: Ie618f20e0c576af65b2eeaf9ef838029cfadf64d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1422702
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Matt Delco <delco@chromium.org>
Reviewed-by: Matt Delco <delco@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>