chromeos-base/hardware_verifier: Add initial cros_workon ebuild
Hardware verifier is a program which compares the output from
`runtime_probe` with a hardware verification spec to verify the
device status.  The source code is located at `platform2`.  The
program is expected to be installed into rootfs when it's ready.

The ebuild this CL contains won't install any file.  Instead,
the ebuild currently only provides a way to compile the code
and run the unit tests.

BUG=chromium:926825
TEST=`FEATURES=test emerge-eve chromeos-base/hardware_verifier`

Cq-Depend: chromium:1445857
Change-Id: I99affd14f3b2ad1e51c836f44a9781edfcb8318d
Reviewed-on: https://chromium-review.googlesource.com/1446011
Commit-Ready: Yong Hong <yhong@google.com>
Tested-by: Yong Hong <yhong@google.com>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Hong-Min Chu <hmchu@chromium.org>