journald: install files under /usr
This brings the journald install in line with most other chromeos
executables. It also puts journalctl in the same place as it is on
Laikitu. This allows us to use the same autotest scripts to gather logs.

The split-usr config option is used to tell systemd that it's being
installed on a system without the /usr merge. The original gentoo
systemd ebuild used the split-usr use flag to control both the
rootprefix and the split-usr config options, so that if you install on a
split-usr system, the root prefix is set to "/" instead of "/usr/". Our
system is a split-usr system, but we want journald to be installed in
/usr/.

BUG=b:120686750
TEST=Ensure system boots and run journalctl

Change-Id: I3d9b83f925611ff6f470a22455d2ce446dbd76ba
Reviewed-on: https://chromium-review.googlesource.com/1431254
Commit-Ready: Christopher Morin <cmtm@google.com>
Tested-by: Christopher Morin <cmtm@google.com>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>