Reland "dev-lang/rust: upgrade to rust 1.32"
This reverts commit 445bbed9191ee6ec51969bc6b44613f7c8e5c3b6.

This includes a revert in the llvm submodule of rust 1.32 that prevents
LLVM from searching for /usr/include/demangle.h, which is not installed
by the sdk builders.

This change also includes a note in UPGRADE.md for upgrading
package.provided and the virtio/rust package.

BUG=chromium:924647
TEST=cros tryjob chromiumos-sdk-tryjob

Change-Id: I054c779a8bad572001005114c19c19434339ff8b
Reviewed-on: https://chromium-review.googlesource.com/1432733
Commit-Ready: Zach Reizner <zachr@chromium.org>
Tested-by: Zach Reizner <zachr@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>
Reviewed-by: Zach Reizner <zachr@chromium.org>