rust, crosvm: Support ignoring path dependencies in patch.crates-io
This prevents out of sandbox accessing while building.

We will use this to discard parts of [patch.crates-io] which should
apply to local developer builds but not to ebuilds.

This CL introduces a replacement of:

    audio_streams = { path = "../../third_party/adhd/audio_streams" } # ignored by ebuild

with empty line while building with emerge.

BUG=chromium:916921
TEST=emerge-kevin crosvm
CQ-DEPEND=CL:1398641

Change-Id: Iec1b130ba1ed710753122d3fb9aab8a97eb3f504
Reviewed-on: https://chromium-review.googlesource.com/1433794
Commit-Ready: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Chih-Yang Hsia <paulhsia@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>