Enable BFQ IO scheduler where supported
The BFQ IO scheduler has much better performance under stress than
the default IO scheduler.

The rule checks if bfq is supported before trying to enable it.
If bfq is not supported (blk-mq is disabled or bfq support is
not built into the kernel), the default IO scheduler is used.

BUG=chromium:902929
TEST=Boot with BFQ enabled in the kernel configuration;
	cross-check by booting on system with BFQ disabled.

Change-Id: I0ebf1ea3dd09b07e42e69eb512fd74003c755e46
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1440381
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>