chromeos-base: Rename IO scheduler udev rule file
The original intent was to use BFQ only for non-rotational devices.
In that context, it made sense to have 'ssd' in the udev rule file name.
However, the rule was updated to apply to all IO devices supporting
bfq, so the original name does not make much sense anymore and is
misleading.

Also update the comment in the udev rule to indicate that we are setting
the active IO scheduler, not the default IO scheduler as originally
suggested.

The actual udev rule remains unchanged.

BUG=chromium:902929
TEST=Boot with BFQ enabled in the kernel configuration;
     cross-check by booting on system with BFQ disabled.

Change-Id: I03a073fa03221abba1f61636d45dba255cb5a4e9
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1450271
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>