cros-kernel2.eclass: Install gdb scripts with KGDB
If we're using kgdb, we probably want to use the gdb scripts that the
kernel ships to get some more debug info. Let's install the scripts into
the image next to the vmlinux so that debugging is a little easier.
Note, to use this you'll need to add an auto-load-safe-path into the
users ~/.gdbinit file, like:

	add-auto-load-safe-path /build/$BOARD/usr/lib/debug/boot/

BUG=None
TEST=echo g > /proc/sysrq-trigger, attach gdb and try lx-cpus, lx-dmesg,
etc.

Change-Id: I94161eb5c6ae90cc8052598436017ab6f36f9ba7
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1453496
Reviewed-by: Douglas Anderson <dianders@chromium.org>