python2: use PGO on the interpreter
This CL turns on PGO for all builds of Python's interpreter.

Python appears to have support for some of this on its own, though:
- that support appears to be confined to GCC in Python2 (Python3 grew
  support for clang at some point)
- for reasons detailed below, we don't want to generate a new profile on
  every `emerge` of Python

...So none of that is really useful to us in Python2.

PGO increases the installed tree size by 8KB, so size isn't an issue
here.

Experimentally, PGO makes Python-heavy operations (such as `equery b
/path/to/foo`) consume 8% fewer user cycles than builds without PGO.

Due to issues with LTO, a comment was added about it, but it wasn't
enabled.

This also includes a script that can be used to generate new profiles.
Since profile generation is nondeterministic, and since it takes
O(minutes), we don't regenerate profiles on each run of the ebuild.

Given the scary comment at the top of python's ebuild, I added a nop
change to dev-install so it's hopefully uprevved.

BUG=None
TEST=emerged the PGO'ed python, built {falco,bob,daisy} from the ground
     up with it.

Change-Id: Idfaf79099f72ad82b0c1c092481da62793adb21d
Reviewed-on: https://chromium-review.googlesource.com/1455417
Commit-Ready: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>