chromeos-bootimage: use altfw payload hashing for all
The primary purpose of this change is to add the hash for the diag
payload.  The only remaining payload that isn't hashed is seabios, so I
added that for the sake of consistency (under the assumption that just
because a hash is there it doesn't mean it has to be used, and seabios
is now opt-in on master so it shouldn't slow down the common case).

This change also changes the name of the hash file from
altfw_hash/${name}.sha256 to altfw/${name}.sha256 to make it easier for
the firmware to derive the hash file name from the image name (by merely
appending .sha256).

BUG=b:124358784
TEST=USE="tianocore seabios diag_payload" emerge-sarien depthcharge
    coreboot chromeos-bootimage
cbfstool /build/sarien/firmware/image-sarien.bin print -r FW_MAIN_A |
    grep -i altfw
altfw/tianocore.sha256    0x8840     raw                32 none
altfw/seabios.sha256      0x88c0     raw                32 none
altfw/diag.sha256         0x8940     raw                32 none

Change-Id: I59f8e590fda8f453e00a80aa402808923761c7b5
Signed-off-by: Matt Delco <delco@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1470548
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Matt Delco <delco@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>