chromeos-base/arc-keymaster: Taint OpenSSL and BoringSSL headers so they don't mix.
Using OpenSSL and BoringSSL in the same binary is dangerous. One source
of problems is that a shared object may unknowingly include headers from
both libraries, potentially leading to undefined behavior at runtime.

This CL taints the headers of OpenSSL in the system and of BoringSSL in
arc-keymaster to safeguard against this problem.

BUG=chromium:733477,b:110136496
TEST=emerge-eve chromeos-base/arc-keymaster

Change-Id: Iba65ec98a11be0314eba142954dd6c09d1ef2d95
Reviewed-on: https://chromium-review.googlesource.com/1464402
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Edman Anjos <edman@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>