ps8751: add firmware blob version 0x3d (61)
this adds new ps8751 TCPC to the build.  this is a test version from
parade with the "continuous alert" issue fixed.  the "-61" package
needs to be unmasked explicitly to use this new blob, so by itself
adding this patch does not affect the build.

BUG=b:123727520
BRANCH=none
TEST=in combination with other paches, updated atlas and verified
	watchdog loop issue is resolved

Change-Id: I38a54256170d2ae29c7c96da98e0ff665b3d80f7
Reviewed-on: https://chromium-review.googlesource.com/1464944
Commit-Ready: Caveh Jalali <caveh@google.com>
Tested-by: Caveh Jalali <caveh@google.com>
Reviewed-by: Matthew Blecker <matthewb@chromium.org>