media-sound: Import alsa-utils from portage-stable
Move alsa-utils temporarily from portage-stable, so that we
can include a patch to aid in ARC++ MIDI testing. This patch has been
accepted upstream, but will only be included in the next alsa-utils
release:

http://git.alsa-project.org/?p=alsa-utils.git;a=commit;h=52c9cc9571659377800fea026cf31b24e2f31394

Once the new alsa-utils is released, we should delete this folder and
revert to using the portage-stable variant.

This CLs retains *all* of the files from the portage-stable alsa-utils
directory except:
- alsa-utils-1.1.6-r1.ebuild (PATCHES variable added).
- The patch file (included in the files/ subdir).

BUG=b:122860765
TEST=emerge-<board> alsa-utils and check that arecordmidi has the new
     functionality after deploying on device.

Change-Id: Idbfcd3bf5819f86b225478cfa612a72ffd1da6b5
Reviewed-on: https://chromium-review.googlesource.com/1472217
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Dan Erat <derat@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>