chromeos-base/arc-keymaster: Setup libhardware and shared libs
Use libhardware headers from arc-toolchain-p and install shared
libraries in the system.

Keymaster depends on libhardware, which is part of the arc toolchain.
We can't inherit from arc-build because we don't want the full
toolchain. This builds a CrOS daemon, and the toolchain conflicts with
the CrOS build tools. Instead we copy the relevant headers from
libhardware to WORKDIR.

We use arc-toolchain-p specifically because this package doesn't support
ARC N, and we can rely on the keymaster backwards compatibility layer
for this to work on P and newer ARC versions.

BUG=b:110136496
TEST=emerge-eve chromeos-base/arc-keymaster

Change-Id: I2654f19f8a941a862c93b18c23dcc0478c36d54a
Reviewed-on: https://chromium-review.googlesource.com/1477239
Commit-Ready: Edman Anjos <edman@chromium.org>
Tested-by: Edman Anjos <edman@chromium.org>
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>