grpc: Fix vsock patch
The vsock code in the kernel checks that an undocumented `svm_zero`
field in the sockaddr_vm struct is always set to 0.  To do this, we have
to allocated the sockaddr_vm struct with gpr_zalloc, which will zero out
the entire structure.

Fix the vsock patch to do that.

BUG=chromium:933513
TEST=vmlog_forwarder starts successfully

Change-Id: I6a017bcc76db24622e23aaa4e604caad23639932
Reviewed-on: https://chromium-review.googlesource.com/1478233
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Stephen Barber <smbarber@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>