test-dlc: a DLC for tast test
This ebuild installs a table file used for DLC dm-verity setup into rootfs.
This ebuild installs a manifest file for imageloader to verify
image content into rootfs.
This ebuild also installs a payload file to stateful partition used by
platform.DLCService tast test.

The exact process to reproduce the table/manifest is described in comments.
The reason that we're not scripting genenating the table file is that a
payload file used in our tast can not be generated using a script since
paygen work is on-going. So we have to pre-populate both the payload file
and table file and rely on updating them manually together.

This package will be included in target-chromium-os behind dlc-test USE
flag.

BUG=chromium:922231
TEST=tast run <IP> platform.DLCService

Change-Id: I40d4b80a587a588acfb4cf8220b7f6305d7a8d43
Reviewed-on: https://chromium-review.googlesource.com/1457397
Commit-Ready: Xiaochu Liu <xiaochu@chromium.org>
Tested-by: Xiaochu Liu <xiaochu@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>