FROMLIST: chromiumos-overlay: import ath10k fimrware 10.2.4.70.70
Import ath10k firmware 10.2.4.70.70 until it is available on
linux-firmware.

Here are the change lists between 70.67 and 70.70 (from QCA):
-------------------------------------------------------------
CL:6552430
Description:Fix for CR 2366654 - Tracker issue 117586409
[Arkham][GES-PMF] AP stop beaconing after successful association
using PMF

Merging the fix from Peregrine Mainline:
CL : 3518904
Issue : SA query action frames are not sent out by firmware in PMF
enabled case after a hard reboot scenario run in MBP.
Root Cause:
When MBP comes out of Power save it sends a Auth Frame.Followed by
which it sends SA query action frame. But as the Client is in power
save it got queued and waiting for client to come out of powersave.
AP is sending SA query at the interval of 200msecs. FW receives 5 SA
query request, but none of the frames are seen over AIR.
When rootcaused,the TID flags status was in paused state.

Fix is to Move the STA to Unpause state once the Auth frame received in
powersave module. Once the state is changed to un-paused, we will be able
to send the SA queries in AIR

BUG=b:117586409
TEST=1) Successfull compilation
     2) Verified md5sum of firmware

Change-Id: I0ddf1694fea1a7e783c11a4ed51ca5dd1781d24b
Reviewed-on: https://chromium-review.googlesource.com/1455416
Commit-Ready: YA-CHIEH WU <ycwu@google.com>
Tested-by: NARAYANRADDI MASTI <nmasti@google.com>
Reviewed-by: Srinivasa duvvuri <sduvvuri@google.com>