dev-install: symlink the entire /etc/portage tree
The only thing that installs into /etc/portage on a device is us, so
rather than create a bunch of dirs and symlinks that mirror the paths
under /usr/share/dev-install/portage/, symlink the directory itself.
This makes understanding the system on the device itself easier.

BUG=None
TEST=precq passes

Change-Id: I9e5e64a1642e0692cdf82714cb07845ec22ff528
Reviewed-on: https://chromium-review.googlesource.com/1494154
Commit-Ready: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Ned Nguyen <nednguyen@google.com>