sys_util: Don't run fork tests under ASAN
Forking in multi-threaded context does not work well with ASAN.  It
especially doesn't work well when the fork is done via a direct clone(2)
system call.  Since the rust test runner spawns a new thread to actually
run the tests, the ASAN runtime in the child process thinks that the
memory owned by the test runner thread is not freed and reports memory
leaks.   This is because clone(2) creates a replica of the calling
thread and its entire address space but not any other threads that might
have been in the parent process.

While this is technically a real leak, it's not actually very useful to
us so just don't run these tests when building with ASAN.

BUG=chromium:935141
TEST='FEATURES=test emerge-$BOARD sys_util` works with and without
     USE=asan

Change-Id: I396944aad27f455594e52a0b4b7c17605ddbacbc
Reviewed-on: https://chromium-review.googlesource.com/1490694
Commit-Ready: Chirantan Ekbote <chirantan@chromium.org>
Tested-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>