linux-firmware: Rename qca-bt USE flag to qca6174a-5-bt
With USE=qca-bt everything in qca/* is installed. Currently the
directory only contains firmware for the QCA6174A-5 (PCIE+USB),
however images for other Qualcomm controllers will be added. Rename
the qca-bt USE flag to qca6174a-5-bt and limit it to the binaries
for the QCA6174A-5.

BUG=b:126438697
TEST=USE=linux_firmware_qca6174a-5-bt emerge-scarlet linux-firmware
  ls /build/scarlet/lib/firmware/qca/
    nvm_usb_00000302.bin  rampatch_usb_00000302.bin
CQ-DEPEND=CL:1492651

Change-Id: If804689dde628a6bdc99e61081070fba8f0dcb41
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1492571
Reviewed-by: Brian Norris <briannorris@chromium.org>