perf: fix hugepage text section problem.
Previous CL crrev.com/c/390488 has a bug when fixing the problem
of annoymous private huge pages. The major problem is that the
previous CL assumes second_event->mmap2.len and
third_event->mmap2.len store ending address of the mmap, and use
the same assumption for event->mmap2.len. But event->mmap2.len is
calculated in original Linux perf code as the length of mmap2.
It results in the calculation of condition are incorrect.
This CL fixes the definitions of second_event->mmap2.len and
third_event->mmap2.len to match the event->mmap2.len, i.e. storing
the length of each mmap2, to avoid confusions in the future.

This CL also improves the condition checking as it checks both
contiguous on virtual address and page offset.

BUG=chromium:931465
TEST=Pass local test on caroline
Without this CL, running
$ perf record -a -e instructions -o /tmp/perf.data /opt/google/chrome/chrome --no-sandbox,
$ perf report -D -i /tmp/perf.data | grep MMAP2 | grep chrome/chrome
gives outputs like:
0 0 0x285e0 [0x80]: PERF_RECORD_MMAP2 7198/7198: [0x55555650a000(0x93dd000) @ 0xfb6000 b3:03 61601 3904683760304533048]: r-xp /opt/google/chrome/chrome
0 0 0x2bf60 [0x80]: PERF_RECORD_MMAP2 7310/7310: [0x55555650a000(0xf6000) @ 0xfb6000 b3:03 61601 3904683760304533048]: r-xp /opt/google/chrome/chrome
0 0 0x2c048 [0x80]: PERF_RECORD_MMAP2 7310/7310: [0x555558400000(0x74e7000) @ 0x2eac000 b3:03 61601 3904683760304533048]: r-xp /opt/google/chrome/chrome
0 0 0x2e388 [0x80]: PERF_RECORD_MMAP2 7323/7323: [0x55555650a000(0xf6000) @ 0xfb6000 b3:03 61601 3904683760304533048]: r-xp /opt/google/chrome/chrome
0 0 0x2e470 [0x80]: PERF_RECORD_MMAP2 7323/7323: [0x555558400000(0x74e7000) @ 0x2eac000 b3:03 61601 3904683760304533048]: r-xp /opt/google/chrome/chrome
...
where the mmaps in pid 7310 and mmaps in pid 7323 are not merged.

With this CL, the output becomes
0 0 0x28180 [0x80]: PERF_RECORD_MMAP2 7198/7198: [0x55555650a000(0x93dd000) @ 0xfb6000 b3:03 61601 3904683760304533048]: r-xp /opt/google/chrome/chrome
0 0 0x2bb00 [0x80]: PERF_RECORD_MMAP2 7310/7310: [0x55555650a000(0x93dd000) @ 0xfb6000 b3:03 61601 3904683760304533048]: r-xp /opt/google/chrome/chrome
0 0 0x2de40 [0x80]: PERF_RECORD_MMAP2 7323/7323: [0x55555650a000(0x93dd000) @ 0xfb6000 b3:03 61601 3904683760304533048]: r-xp /opt/google/chrome/chrome
...

perf.data generated with crosperf tool shows similar results.

Change-Id: Icdb7dd8747e4a4f54ba9ea0d5bb1a9216a0453e0
Reviewed-on: https://chromium-review.googlesource.com/1496075
Commit-Ready: Tiancong Wang <tcwang@google.com>
Tested-by: Tiancong Wang <tcwang@google.com>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>