chromiumos-overlay: Add grpc-web, allows browser to access grpc
This ebuild does not exist in gentoo so it is hand crafted.

It generates a plugin for protoc, which will be used at
build time, so this is not tested on board specific basis.

I will follow this with a CL to add this to the sdk.

TEST=sudo emerge net-libs/grpc-web, trybot run with a simple compile.
BUG=chromium:682415

Change-Id: I47be7a68b14f6b98b67a6abc3ed73505c3caf4b8
Reviewed-on: https://chromium-review.googlesource.com/1504813
Commit-Ready: Keith Haddow <haddowk@chromium.org>
Tested-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Jason Clinton <jclinton@chromium.org>