crosvm: Compile with TPM when USE=tpm2 is set
This effectively enables crosvm software TPM support for most boards.

Note that this changes the name of the relevant USE flag from crosvm-tpm
to the Chrome OS-wide tpm2 flag.  This is because we now DEPEND on
trunks, which has a blocker for !app-crypt/tpm-tools, and this package
is pulled in on boards that set USE=tpm (rather than USE=tpm2).

The TPM device still requires `--software-tpm` flag passed to the crosvm
invocation in order to give the VM a TPM, so what this CL is changing is
that the TPM implementation is compiled in, not that every VM gets a TPM
by default.

BUG=chromium:911799
TEST=emerge-nami crosvm

Change-Id: I99192467e72f197f739a17d50b5759e74ef0fe4e
Signed-off-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1507248
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: Zach Reizner <zachr@chromium.org>