chromeos-cr50-scripts: add cr50-set-sn-bits.sh
This script will write SN Bits in INFO1 through gsctool. It is used by
the factory process.

For now the new script is only included if the zero-touch USE flag is
defined for a board. gooftool will not attempt to run this script if
it is missing.

BUG=chromium:939223
TEST=manual

CQ-DEPEND=CL:1514215
CQ-DEPEND=CL:1519567

Change-Id: Ife4c3aaf7cdd28d7ed5e47e288d5ead180494e09
Reviewed-on: https://chromium-review.googlesource.com/1507327
Commit-Ready: Yves Arrouye <drcrash@chromium.org>
Tested-by: Yves Arrouye <drcrash@chromium.org>
Reviewed-by: Andrey Pronin <apronin@chromium.org>