libva-amdgpu-driver: Correct libva version in RDPEND
libva doesn't have a backward compatibility between 1.8.3 and 2.4.0.
Since libva-amdgpu-driver ebuild is compilable against libva 2.4.0, we can
update the required version to libva 2.4.0. Thanks to this, a reader doen't have
to have a coneren that libva-amdgpu-driver is only compilable with 1.8.3, but
not 2.4.0.

BUG=None
TEST=emerge-grunt libva-amdgpu-driver with CL:1464335
CQ-DEPEND=CL:1464335

Change-Id: Ie9e57165d89c0c4a1da2925fe2652327fef36367
Reviewed-on: https://chromium-review.googlesource.com/1507354
Commit-Ready: Hirokazu Honda <hiroh@chromium.org>
Tested-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>