cros-kernel2.eclass: Don't inherit linux-info
There appears to be no valid reason to inherit linux-info.  Shave a
tiny bit of time off the kernel build.

This is minor enough that any difference it makes to the build time is
in the noise and hard to measure, but maybe we still want it just to
clean things up?

BUG=chromium:767073
TEST=Build still works

Change-Id: I0d213b5a0618d841858c3fe8ec024ba42b0dab90
Reviewed-on: https://chromium-review.googlesource.com/1509615
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>