[compiler_wrapper] Add a double_exec utility with tests
This function is pretty simple in spirit: it invokes the compiler once.
If that fails and it appears to be due to -Werror, it invokes the
compiler again. If the second run succeeds, we pretend the first run
never happened. If the first run succeeds, we pretend the second run
never happened.

In either case, if we think that -Werror went into failing a build,
we'll log a bit of JSON into a well-known /tmp directory. The intent is
for a script on a buildbot to pick up on all of the files in that /tmp
directory and postprocess them.

BUG=chromium:914081
TEST=Ran the new test; chromiumos-sdk tryjob

Change-Id: Iea79300f37b8de2d2648bab2ff343ded52f6c422
Reviewed-on: https://chromium-review.googlesource.com/1509716
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: George Burgess <gbiv@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>