cros-kernel2.eclass: Avoid calling olddefconfig when possible
Calling 'kmake olddefconfig' adds 1.5 seconds to an incremental build.
Most of the time this is totally unnecessary since we haven't touched
the config nor any Kconfig files.  The worst part about this 1.5
seconds is that it happens _before_ the source code gets compiled, so
if you're just trying to figure out compile errors you need to wait
this 1.5 seconds over and over again.

Let's save this 1.5 seconds most times by keeping a copy of both the
defconfig and the raw output of cat-ing all of our config fragments.
If the raw output is the same then we can restore the defconfig from
last time and save the time.

Checking the raw cat of the fragments will catch cases where the
chromeos/config files or the fragments used from cros-kernel2.eclass
changed.  ...but if the Kconfigs in the kernel themselves changed then
we'll end up calling into the kernel with a non-defconfig.  This is
mostly OK--the kernel will detect it, fall to a slow path, and (very
noisily) make a defconfig for you.  ...but to be totally paranoid
we'll actually redo the olddefconfig and kernel compile if we detect
this.  This makes the Kconfig changed case slower than before this
patch but that's likely worth it.

Timings from my workstation building:
  FEATURES="-buildpkg nostrip" \
  USE="debugobjects lockdebug klang kgdb vtconsole" \
  emerge-cheza --nodeps chromeos-kernel-4_19

Before this change, a no-op build:
  0m32.896s, 0m33.223s, 0m33.335s

Before this change, a no-op build simulating a Kconfig change (by
commenting out the "touch -r "${temp_config}" "$(get_build_cfg)" in
the eclass):
  0m38.717s, 0m38.193s, 0m38.113s

After this change, a no-op build:
  0m31.304s, 0m31.776s, 0m31.869s

After this change, a no-op build simulating a Kconfig change by doing
a "cp cros-old-config cros-old-defconfig" between builds:
  0m38.185s, 0m37.868s, 0m38.477s

Thus we get a ~1.5 second savings most times.  If we ever "get it
wrong" and end up in the pathological case we'll actually end up
churning an extra ~13 seconds that we didn't have to, but that should
be highly unusual and at least we'll still be correct.

BUG=chromium:767073
TEST=See above timings; build still seems working

Change-Id: I7e2ef50b8ac838fb430b133aad5c3152431ae9f4
Reviewed-on: https://chromium-review.googlesource.com/1516236
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>