crosvm: Remove manifest now that we use cros-rust eclass
The Manifest includes hashes of all the entries in the SRC_URI field
from the ebuild. All of crosvm's dependency crates used to be in that
list, but now that crosvm has moved to the cros-rust eclass, SRC_URI is
empty so there is no longer a need for the Manifest file.

TEST=ebuild crosvm-9999.ebuild manifest

Change-Id: I06c08508dcd1ff61227b6df4cee21aed786d62fa
Reviewed-on: https://chromium-review.googlesource.com/1516004
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: David Tolnay <dtolnay@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>