sys-devel/arc-llvm: Copy from llvm-7.0.1
This copies the current llvm-7.0.1 ebuild from the sys-devel/llvm
package without changes so that the resulting diffs are easier to
review.

Because this is copying the existing state, this change doesn't build;
another commit will follow to make a buildable package.

BUG=b:128363272
TEST=none at this point
CQ-DEPEND=CL:1521845

Change-Id: I28658b3309b77c2cc73e3e031059b0a9b2bdbc45
Reviewed-on: https://chromium-review.googlesource.com/1521842
Commit-Ready: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Chad Versace <chadversary@chromium.org>
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>