cros-kernel2.eclass: Get ready to kill firmware_install kernel USE flag
As per CL:729700, ("linux-firmware: Clean linux-firmware (removing
most of the files)"):

> "make firmware_install" was probably the wrong thing to be doing all
> along.  Running "make firmware_install" installs _every single_
> firmware file even if there is no driver enabled for it (!).  If we
> only want firmware files that are actually needed then "make
> modules_install" is the right thing to do.

As of that same CL (now 1.5 years old), all known users of the
firmware_install flag are gone.  ...but it seems we forgot to take
support for it out of the eclass.

Let's add a period where setting the firmware_install USE flag counts
as a build error.  After this has been in the build for a while, we
can fully delete it.  This seems better than deleting it right away in
case some non-public overlay (not that I can find any) is still
referring to it.

BUG=chromium:767073, chromium:770230
TEST=Everything still builds

Change-Id: I1f46caaeac68f1ea40b94a8335c4e7595e031922
Reviewed-on: https://chromium-review.googlesource.com/1518795
Commit-Ready: Douglas Anderson <dianders@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>